Code cleanup. - separated hotplug functionality from device driver for USB devices - factored out common code to minimize redundancy for different platforms (linux/freebsd) - created an internal "device" data type to allow handling both USB and serial devices   transparently without conditions everywhere.
Also fixed a minor performance bug. If there is no device plugged in, the repeat interval is ignored.


git-svn-id: https://spacenav.svn.sourceforge.net/svnroot/spacenav@139 9acbd7d0-4e34-0410-9de4-80efa1d9a8fd
