+++ trunk/spacenavd/src/dev.h
@@ -1,6 +1,6 @@
 /*
 spacenavd - a free software replacement driver for 6dof space-mice.
-Copyright (C) 2007-2010 John Tsiombikas <nuclear@member.fsf.org>
+Copyright (C) 2007-2012 John Tsiombikas <nuclear@member.fsf.org>
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -15,28 +15,32 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-
-#ifndef DEV_H_
-#define DEV_H_
+#ifndef SPNAV_DEV_H_
+#define SPNAV_DEV_H_
 
 #include "config.h"
-#include "event.h"
 
-/* device hotplug detection */
-int init_hotplug(void);
-void shutdown_hotplug(void);
-int get_hotplug_fd(void);
+struct dev_input;
 
-int handle_hotplug(void);
+#define MAX_DEV_NAME	256
 
+struct device {
+	int fd;
+	void *data;
+	char name[MAX_DEV_NAME];
 
-/* device handling */
+	void (*close)(struct device*);
+	int (*read)(struct device*, struct dev_input*);
+	void (*set_led)(struct device*, int);
+};
+
 int init_dev(void);
 void shutdown_dev(void);
 int get_dev_fd(void);
+#define is_dev_valid()	(get_dev_fd() >= 0)
 
 int read_dev(struct dev_input *inp);
 
 void set_led(int state);
 
-#endif	/* DEV_H_ */
+#endif	/* SPNAV_DEV_H_ */
