+++ trunk/spacenavd/src/spnavd.c
@@ -1,6 +1,6 @@
 /*
 spacenavd - a free software replacement driver for 6dof space-mice.
-Copyright (C) 2007-2010 John Tsiombikas <nuclear@member.fsf.org>
+Copyright (C) 2007-2012 John Tsiombikas <nuclear@member.fsf.org>
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -26,6 +26,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include <sys/select.h>
 #include "spnavd.h"
 #include "dev.h"
+#include "hotplug.h"
 #include "client.h"
 #include "proto_unix.h"
 #ifdef USE_X11
@@ -144,7 +145,7 @@ int main(int argc, char **argv)
 
 		do {
 			struct timeval tv, *timeout = 0;
-			if(cfg.repeat_msec >= 0 && !in_deadzone()) {
+			if(is_dev_valid() && cfg.repeat_msec >= 0 && !in_deadzone()) {
 				tv.tv_sec = cfg.repeat_msec / 1000;
 				tv.tv_usec = cfg.repeat_msec % 1000;
 				timeout = &tv;
